pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-server-sentinel"
version = "0.1.0"
description = "Microsoft Sentinel MCP Server implementation"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    {name = "Daniel Streefkerk", email = "email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
]
dependencies = [
    "azure-identity>=1.21.0",
    "azure-mgmt-authorization>=4.0.0",
    "azure-mgmt-loganalytics>=12.0.0",
    "azure-mgmt-securityinsight==2.0.0b2",
    "azure-monitor-query>=1.4.1",
    "cachetools>=5.5.2",
    "clr>=1.0.3",
    "dotenv>=0.9.9",
    "mcp[cli]>=1.6.0",
    "pythonnet>=3.0.1",
    "setuptools>=79.0.0",
]

[tool.setuptools]
packages = {find = {}}
include-package-data = true

[project.optional-dependencies]
dev = [
    "autopep8>=2.3.2",
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.4.4",
    "coverage>=7.0.0",
]


[project.scripts]
mcp-server-sentinel = "main:main"
download-kusto-dll = "download_dll:download_kusto_dll"
_mcp_sentinel_post_install = "post_install:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/mcp-server-sentinel"
"Bug Tracker" = "https://github.com/yourusername/mcp-server-sentinel/issues"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
ignore = []
